Rental package jack

+++ code/src/main/java/org/example/rental/RentalOrder.java
@@ -57,7 +57,7 @@ public class RentalOrder implements IRentalOrder {
     }
 
     @Override
-    public double getFee() {
+    public float getFee() {
         return fee;
     }
